requirements.txt
streamlit
pandas
numpy
scikit-learn
tensorflow-cpu==2.15.0
tmdbv3api
requests
joblib
